Refactor: extract validation schema
